renamed settings to handler_settings and fixed imports
